+++ src/components/Table/tableChangelogData.ts
@@ -10,6 +10,17 @@ import { ChangelogData } from "../../utils/ComponentChangelogTable";
 
 export const changelogData: ChangelogData[] = [
   {
+    date: "Prerelease",
+    version: "Prerelease",
+    type: "Update",
+    affects: ["Accessibility", "Documentation", "Functionality", "Styles"],
+    notes: [
+      "Added the `isScrollable` prop to enable horizontal scrolling.",
+      "Added the `customStyles` prop to allow for custom column styles (i.e. width, text alignment, etc.).",
+      "Added the `tableTextSize` prop to set the size of the text within the table.",
+    ],
+  },
+  {
     date: "2024-05-09",
     version: "3.1.2",
     type: "Update",